requirements.txt
Package           Version
----------------- -----------
attrs             23.2.0
black             24.2.0
certifi           2024.2.2
click             8.1.7
exceptiongroup    1.2.0
h11               0.14.0
idna              3.6
iniconfig         2.0.0
mypy-extensions   1.0.0
outcome           1.3.0.post0
packaging         23.2
pathspec          0.12.1
pip               22.2.2
platformdirs      4.2.0
pluggy            1.4.0
PySocks           1.7.1
pytest            8.0.1
python-dotenv     1.0.1
selenium          4.17.2
setuptools        65.3.0
sniffio           1.3.0
sortedcontainers  2.4.0
tomli             2.0.1
trio              0.24.0
trio-websocket    0.11.1
typing_extensions 4.9.0
urllib3           2.2.1
wheel             0.37.1
wsproto           1.2.0
